rfid/requirements.txt
mfrc522
spidev
RPi.GPIO